requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
py4j==0.10.9.9
pyspark==4.1.1
PyYAML==6.0.3
requests==2.32.5
urllib3==2.6.3
